VFB-6: some refactoring
